+++ generator/config-validator.js
