+++ generator/logger.js
