+++ generator/schema/chain.schema.js
