+++ generator/schema/tee.schema.js
